refactor(pdfjson): extract text element cursor
