+++ app/core/src/main/java/stirling/software/SPDF/service/PdfJsonConversionService.java
@@ -115,6 +115,7 @@ import stirling.software.SPDF.model.json.PdfJsonStream;
 import stirling.software.SPDF.model.json.PdfJsonTextColor;
 import stirling.software.SPDF.model.json.PdfJsonTextElement;
 import stirling.software.SPDF.service.pdfjson.PdfJsonFontService;
+import stirling.software.SPDF.service.pdfjson.PdfTextElementCursor;
 import stirling.software.SPDF.service.pdfjson.PdfTextMergeHelper;
 import stirling.software.SPDF.service.pdfjson.encoding.PdfTextEncoder;
 import stirling.software.SPDF.service.pdfjson.font.PdfFontResolver;
@@ -3919,7 +3920,7 @@ public class PdfJsonConversionService {
             PDFStreamParser parser = new PDFStreamParser(page);
             List<Object> tokens = parser.parse();
             log.debug("Parsed {} tokens for rewrite", tokens.size());
-            TextElementCursor cursor = new TextElementCursor(elements);
+            PdfTextElementCursor cursor = new PdfTextElementCursor(elements);
             PDFont currentFont = null;
             String currentFontName = null;
             PdfJsonFont currentFontModel = null;
@@ -4033,7 +4034,7 @@ public class PdfJsonConversionService {
             PDFont font,
             PdfJsonFont fontModel,
             String expectedFontName,
-            TextElementCursor cursor,
+            PdfTextElementCursor cursor,
             boolean removeOnly)
             throws IOException {
         if (font == null) {
@@ -4047,7 +4048,7 @@ public class PdfJsonConversionService {
         log.trace(
                 "rewriteShowText consuming {} glyphs at cursor index {} for font {}",
                 glyphCount,
-                cursor.index,
+                cursor.getIndex(),
                 expectedFontName);
         List<PdfJsonTextElement> consumed = cursor.consume(expectedFontName, glyphCount);
         if (consumed == null) {
@@ -4090,7 +4091,7 @@ public class PdfJsonConversionService {
             PDFont font,
             PdfJsonFont fontModel,
             String expectedFontName,
-            TextElementCursor cursor,
+            PdfTextElementCursor cursor,
             boolean removeOnly)
             throws IOException {
         if (font == null) {
@@ -4143,76 +4144,6 @@ public class PdfJsonConversionService {
             }
         }
         return true;
-    }
-
-    private static class TextElementCursor {
-        private final List<PdfJsonTextElement> elements;
-        private int index = 0;
-
-        TextElementCursor(List<PdfJsonTextElement> elements) {
-            this.elements = elements;
-        }
-
-        boolean hasRemaining() {
-            return index < elements.size();
-        }
-
-        int remaining() {
-            return Math.max(0, elements.size() - index);
-        }
-
-        List<PdfJsonTextElement> consume(String expectedFontName, int glyphCount) {
-            if (glyphCount <= 0) {
-                return Collections.emptyList();
-            }
-            List<PdfJsonTextElement> consumed = new ArrayList<>();
-            int remaining = glyphCount;
-            while (remaining > 0 && index < elements.size()) {
-                PdfJsonTextElement element = elements.get(index);
-                if (!fontMatches(expectedFontName, element.getFontId())) {
-                    log.debug(
-                            "Cursor consume failed: font mismatch (expected={}, actual={}) at element {}",
-                            expectedFontName,
-                            element.getFontId(),
-                            index);
-                    return null;
-                }
-                consumed.add(element);
-                remaining -= countGlyphs(element);
-                index++;
-            }
-            if (remaining > 0) {
-                log.debug(
-                        "Cursor consume failed: ran out of elements (remaining={}, currentIndex={}, total={})",
-                        remaining,
-                        index,
-                        elements.size());
-                return null;
-            }
-            return consumed;
-        }
-
-        private boolean fontMatches(String expected, String actual) {
-            if (expected == null || expected.isEmpty()) {
-                return true;
-            }
-            if (actual == null) {
-                return false;
-            }
-            return Objects.equals(expected, actual);
-        }
-
-        private int countGlyphs(PdfJsonTextElement element) {
-            int[] codes = element.getCharCodes();
-            if (codes != null && codes.length > 0) {
-                return codes.length;
-            }
-            String text = element.getText();
-            if (text != null && !text.isEmpty()) {
-                return Math.max(1, text.codePointCount(0, text.length()));
-            }
-            return 1;
-        }
     }
 
     private Map<String, PDFont> buildFontMap(
